requirements.txt
absl-py==2.1.0
blinker==1.8.2
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dominate==2.9.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
gmpy2==2.1.2
grpcio==1.63.0
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.4
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pip==24.0
protobuf==5.26.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tifffile==2024.5.10
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
typing_extensions==4.11.0
urllib3==2.1.0
visdom==0.2.4
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
